requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
openpyxl==3.0.9
metagpt==0.1
tavily-python==0.1.8
python-dotenv==0.19.0 